Fixed a typo in the class name.
